requirements.txt
# requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
aiohttp==3.9.5 
google-api-python-client==2.133.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pytz==2024.1